GPU: Move things around a little, out of DrawEngine.
