+++ GPU/Common/DrawEngineCommon.h
@@ -88,8 +88,6 @@ public:
 
 	bool GetCurrentSimpleVertices(int count, std::vector<GPUDebugVertex> &vertices, std::vector<u16> &indices);
 
-	static u32 NormalizeVertices(u8 *outPtr, u8 *bufPtr, const u8 *inPtr, VertexDecoder *dec, int lowerBound, int upperBound, u32 vertType);
-
 	// Dispatches the queued-up draws.
 	virtual void Flush() = 0;
