+++ GPU/Common/SplineCommon.h
@@ -22,21 +22,11 @@
 #include "Common/Swap.h"
 #include "GPU/Math3D.h"
 #include "GPU/ge_constants.h"
+#include "GPU/Common/TransformCommon.h"
+
 #include "Core/Config.h"
 
 #define HALF_CEIL(x) (x + 1) / 2 // Integer ceil = (int)ceil((float)x / 2.0f)
-
-// PSP compatible format so we can use the end of the pipeline in beziers etc
-// 8 + 4 + 12 + 12 = 36 bytes
-struct SimpleVertex {
-	float uv[2];
-	union {
-		u8 color[4];
-		u32_le color_32;
-	};
-	Vec3Packedf nrm;
-	Vec3Packedf pos;
-};
 
 class SimpleBufferManager;
